refactor(retention): relocate rollup SQL adapters to internal/database (WS-B5)
WS-B repository-port discipline: the retention engine
(internal/timeseries/retention) orchestrates rollup + purge cadence
through its consumer-side RollupSource port; the concrete SQL adapters
now live in internal/database where persistence belongs.

Changes:
- Add internal/database/rollup_sources.go: MetricsRollupSource +
  ProbeRollupSource (behavior-preserving relocation; SQL copied verbatim,
  hoursPerDay reused from database package scope)
- Delete internal/timeseries/retention/source_metrics.go and
  source_probe.go (hourFormat / dayFormat / purgeRows move with them)
- Rewire internal/api/server.go initRetentionEngine to call
  database.NewProbeRollupSource / database.NewMetricsRollupSource
- Add timeseries-no-persistence depguard rule (RED-proven before
  relocation, GREEN after) to gate future regressions
